third_party/blink/web_tests/platform/generic/fast/forms/input-type-change-focusout-expected.txt

This tests that changing an input element's type or speech attribute does not cause focusout event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS focusoutCalled is false
PASS focusoutCalled is false
PASS focusoutCalled is false
PASS focusoutCalled is false
PASS focusoutCalled is false
PASS successfullyParsed is true

TEST COMPLETE
